temporary work around for bugzilla being unable to send mail to webteam
The current setup ( 13/02/2011 ) is that most active mls are on
zarb.org, using @mageia domain. But we want to start hosting
aliases on the current ml server ( and main smtp, ie alamut ).
So we need to dispatch old mls to zarb, and the rest on alamut.
Old mls are all prefixed by mageia-*@mageia this hack should
be removed once all mls are migrated from zarb.org

git-svn-id: svn+ssh://svn.mageia.org/svn/adm/puppet@1063 db665093-e049-48a9-b3c2-13f1584ce46d