buildsystem: add option to base a distribution on an other
It is now possible to add the option 'based_on' to a distribution
defined in buildystem::var::distros. This is useful for distributions
like infra_1 or infra_2 which provide a new repository for an existing
distribution. When repositories from an other distribution are defined
in 'based_on', the repositories are linked with a symbolic link, and
defined in media.cfg.

git-svn-id: svn+ssh://svn.mageia.org/svn/adm/puppet@3148 db665093-e049-48a9-b3c2-13f1584ce46d