buildsystem: remove use of .each on hashes
Remove use of "each" methods on hashes in template files. With ruby 1.8,
the order in which the hash is enumerated is random. When using the each
method to scan all hash keys and elements in template files, puppet will
do unnecessary changes in files installed on the server, making it more
difficult to see the real changes. In order to avoid this, we stop using
the "each" method on hashes and instead use the "keys" method to get an
array of all keys and sort it alphabetically.

With ruby 1.9 the hashes enumerate their values in the order that the
corresponding keys were inserted, but we are still using ruby 1.8 with
Mageia 2.

git-svn-id: svn+ssh://svn.mageia.org/svn/adm/puppet@3149 db665093-e049-48a9-b3c2-13f1584ce46d