buildsystem/upload.conf: add arch and mandatory_arch
Add arch and mandatory_arch settings in iurt/upload.conf template.
The arch setting is used by ulri to know the list of architectures for
which the packages need to be built. The mandatory_arch setting is used
by emi to know the list of architectures for which a package needs to be
built successfully before it can be installed in the repository.

git-svn-id: svn+ssh://svn.mageia.org/svn/adm/puppet@3176 db665093-e049-48a9-b3c2-13f1584ce46d