mgagit: Add new templates for generationg per-commit notification emails.
This uses the git multimail project from upstream revision 3fcd7bffef
which is the master revision at the time of writing.

It also includes a monkey-patched version which we will use which adds
the ability to include links to the gitweb/cgit URL and also links
to any supported bugtracker.

git-svn-id: svn+ssh://svn.mageia.org/svn/adm/puppet@3354 db665093-e049-48a9-b3c2-13f1584ce46d