Add rsync access to packages svn repository for lissyx
lissyx is working on a PhD thesis which involves some study of Mageia
releases. In order to do that he needs to do a lot of svn requests on
all packages. To avoid loading our server with those requests, I
proposed him to download a full copy of the svn repository, so he can do
the requests locally without loading our server.

git-svn-id: svn+ssh://svn.mageia.org/svn/adm/puppet@3366 db665093-e049-48a9-b3c2-13f1584ce46d